DTO de productos, ítems, Carrito, checkout y tests

+++ back-ecommerce/src/main/java/com/uade/ecommerce/dto/CheckoutRespuestaDTO.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -12,5 +14,5 @@ import lombok.Setter;
 public class CheckoutRespuestaDTO {
 
     private String mensaje;
-    private Double total;
+    private BigDecimal total;
 }
